Adding the React Routing (#19)
* Adding the React Routing

Adding the code to navigate through different applications of the page. Still need to update how to work with sign Up and login page

* Adding some other pages to the routing

The button event handler to navigate through different pages doesn't work right now. I will keep working on this.
But, as of now all the links to navigate through different pages of the app is working.